wasmExpGo/hestiaGo/hestiaOS/hestiaWASM: added IsValueOK function
Since there is a need to check a return value is convertable back
to Javascript object and the current 'js.ValueOf' actually panics
instead of safely returns an error, I need to facilitate such
support checking table. Hence, let's do this.

This patch adds IsValueOK function to check the value type is
convertable to Javascript object.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>