root: reflect the latest status
Since we now have an overview of CSS, JS, and HTML interactions from
Go WASM, I can conclude the experiement and proceed to write the
research paper out for development. Hence, let's update the README.md
to match the latest findings.

This patch updates the repository to reflect the latest experiment
findings.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>